+++ src/losses/criterion.py
@@ -8,4 +8,4 @@ TverskyLoss = smp.losses.TverskyLoss(mode='multilabel', log_loss=False)
 
 def criterion(pr, gt):
     loss = F.CrossEntropyLoss()
-    return 0.5*loss(pr, gt) + 0.5*TverskyLoss(pr, gt)
+    return 0.5*loss(pr, gt.int()) + 0.5*TverskyLoss(pr, gt)
